requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic>=2.0
pydantic-settings
python-jose[cryptography]
passlib[bcrypt,argon2]
argon2-cffi
python-dotenv
pydantic[email]
python-jose